Split up tools/prompts/resources
